+++ unity-samples/Assets/Scripts/ChangeUser.cs
@@ -1,9 +1,19 @@
 ﻿using Appboy;
+using Appboy.Utilities;
 using System;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 using Utilities;
+using UnityEngine.Networking;
+using System.Text;
+using System.Threading.Tasks;
+
+using System.Collections;
+using System;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.Networking;
 
 public class ChangeUser : MonoBehaviour {
 
@@ -11,12 +21,36 @@ public class ChangeUser : MonoBehaviour {
   
   public void OnDoneButtonClicked() {
     Debug.Log(String.Format("Changing user ID to {0}", UserIDField.text));
-    Appboy.AppboyBinding.ChangeUser(UserIDField.text);
-    SceneManager.LoadScene(Constants.MainMenuScene);
+    getTokenSetAndExit(UserIDField.text);
   }
   
   public void OnCancelButtonClicked() {
     Debug.Log(String.Format(Constants.ChangeUserCancelLog));
     SceneManager.LoadScene(Constants.MainMenuScene);
+  }
+
+  private async void getTokenSetAndExit(string userId) {
+    var request = new UnityWebRequest ("https://us-central1-jwt-responder.cloudfunctions.net/getToken", "POST");
+    byte[] bodyRaw = Encoding.UTF8.GetBytes("{\"data\": {\"user_id\": \"" + userId + "\"}}");
+    request.uploadHandler = (UploadHandler) new UploadHandlerRaw(bodyRaw);
+    request.downloadHandler = (DownloadHandler) new DownloadHandlerBuffer();
+    request.SetRequestHeader("Content-Type", "application/json");
+    request.SetRequestHeader("Accept", "application/json");
+    var operation = request.SendWebRequest();
+
+    while (!operation.isDone) {
+      await Task.Yield();
+    }
+    var jsonResponse = request.downloadHandler.text;
+
+    if (request.error != null) {
+      Debug.Log("Error: " + request.error);
+    } else {
+      Debug.Log("Status Code: " + request.responseCode);
+      JSONClass json = (JSONClass) JSON.Parse(jsonResponse);
+      var token = json["data"]["token"];
+      Appboy.AppboyBinding.ChangeUser(userId, token);
+      SceneManager.LoadScene(Constants.MainMenuScene);  
+    }
   }
 }